Add e2e test for fluxctl install with config file
This adds a smoke test for installing with

    fluxctl install --config-file=...

To be able to do so, the e2e testing scripts have to be rejigged a
little so that arbitrary args for `fluxctl install` can be supplied to
install_flux_with_fluxctl.

The directory structure of kustomizations is altered a little, so that
a kustomization can refer to a file from `fluxctl install` in a
directory under it (Kustomize won't let you refer to a file outside
the directory of the kustomization).