Obviate dependency on helm.chartutil
With a bit of care, it's possible to construct the data section of the
configmap or secret resource within the template (given the file
contents), rather than using chartutil.Files. Prior to pkg/install
being its own module, this wasn't an issue; but now, it's well worth
avoiding the dependency on Helm.